@@ -1,99 +1,126 @@
 {
-  "name": "storyblok-js-client",
-  "version": "5.0.0",
-  "description": "Universal JavaScript SDK for Storyblok's API",
-  "license": "MIT",
-  "types": "./types/index.d.ts",
-  "main": "./dist/index.cjs.js",
-  "module": "./dist/index.es.js",
-  "unpkg": "./dist/index.standalone.js",
-  "jsdelivr": "./dist/index.standalone.js",
-  "source": "source/index.js",
-  "files": [
-    "dist",
-    "source",
-    "types",
-    "test"
-  ],
-  "scripts": {
-    "clean": "rimraf dist/",
-    "lint": "eslint --max-warnings=0 './**/*.{ts,js}'",
-    "prettier": "prettier './**/*.js' './**/*.ts' './**/*.md' --write",
-    "build": "run-s clean tsc build:**",
-    "build:es5": "ENABLE_BABEL=yes rollup --config",
-    "build:es6": "ENABLE_BABEL=yes ES6=yes rollup --config",
-    "build:standalone:es5": "STANDALONE=yes ENABLE_BABEL=yes rollup --config",
-    "build:standalone:es6": "STANDALONE=yes ENABLE_BABEL=yes ES6=yes rollup --config",
-    "test": "NODE_ENV='test' jest",
-    "tsc": "tsc",
-    "prepare": "npm run build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/storyblok/storyblok-js-client.git"
-  },
-  "keywords": [
-    "storyblok"
-  ],
-  "author": "Alexander Feiglstorfer <delooks@gmail.com>",
-  "bugs": {
-    "url": "https://github.com/storyblok/storyblok-js-client/issues"
-  },
-  "homepage": "https://github.com/storyblok/storyblok-js-client#readme",
-  "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/plugin-transform-runtime": "^7.11.5",
-    "@babel/preset-env": "^7.11.5",
-    "@commitlint/cli": "^17.0.2",
-    "@commitlint/config-conventional": "^17.0.2",
-    "@rollup/plugin-babel": "^5.2.1",
-    "@rollup/plugin-commonjs": "^22.0.1",
-    "@rollup/plugin-json": "^5.0.1",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@types/isomorphic-fetch": "^0.0.36",
-    "@typescript-eslint/eslint-plugin": "^5.26.0",
-    "@typescript-eslint/parser": "^5.26.0",
-    "babel-jest": "^26.3.0",
-    "core-js": "^3.6.5",
-    "eslint": "^8.9.0",
-    "eslint-config-prettier": "^8.4.0",
-    "husky": "^8.0.1",
-    "jest": "^29.2.0",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.5.1",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.68.0",
-    "rollup-plugin-mjs-entry": "^0.1.1",
-    "rollup-plugin-terser": "^7.0.2",
-    "typescript": "^4.7.2"
-  },
-  "peerDependencies": {},
-  "dependencies": {
-    "isomorphic-fetch": "^3.0.0"
-  },
-  "release": {
-    "branches": [
-      "master",
-      {
-        "name": "beta",
-        "prerelease": true
-      },
-      {
-        "name": "alpha",
-        "prerelease": true
-      }
-    ]
-  },
-  "commitlint": {
-    "extends": [
-      "@commitlint/config-conventional"
-    ]
-  },
-  "prettier": {
-    "semi": false,
-    "singleQuote": true,
-    "tabWidth": 2,
-    "trailingComma": "es5",
-    "useTabs": true
-  }
+	"name": "storyblok-js-client",
+	"version": "5.0.0",
+	"description": "Universal JavaScript SDK for Storyblok's API",
+	"license": "MIT",
+	"types": "./dist/types/entry.esm.d.ts",
+	"main": "./dist/index.umd.js",
+	"module": "./dist/index.mjs",
+	"unpkg": "./dist/index.umd.js",
+	"jsdelivr": "./dist/index.umd.js",
+	"source": "src/index.ts",
+	"exports": {
+		".": {
+			"import": "./dist/index.mjs",
+			"require": "./dist/index.umd.js"
+		},
+		"./richTextResolver": {
+			"import": "./dist/richTextResolver.mjs",
+			"require": "./dist/richTextResolver.umd.js"
+		},
+		"./schema": {
+			"import": "./dist/schema.mjs",
+			"require": "./dist/schema.umd.js"
+		}
+	},
+	"files": [
+		"dist",
+		"src",
+		"tests"
+	],
+	"scripts": {
+		"lint": "eslint --max-warnings=0 './src/**/*.{ts,js}'",
+		"prettier": "prettier './**/*.js' './src/**/*.ts' './**/*.md' --write",
+		"build": "node vite.build.mjs && tsc",
+		"demo": "vite serve playground",
+		"test": "vitest",
+		"prepare": "npm run build"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/storyblok/storyblok-js-client.git"
+	},
+	"keywords": [
+		"storyblok"
+	],
+	"author": "Alexander Feiglstorfer <delooks@gmail.com>",
+	"bugs": {
+		"url": "https://github.com/storyblok/storyblok-js-client/issues"
+	},
+	"homepage": "https://github.com/storyblok/storyblok-js-client#readme",
+	"devDependencies": {
+		"@commitlint/cli": "^17.2.0",
+		"@commitlint/config-conventional": "^17.2.0",
+		"@tsconfig/recommended": "^1.0.1",
+		"@typescript-eslint/eslint-plugin": "^5.43.0",
+		"@typescript-eslint/parser": "^5.43.0",
+		"eslint": "^8.27.0",
+		"eslint-config-prettier": "^8.5.0",
+		"husky": "^8.0.2",
+		"isomorphic-fetch": "^3.0.0",
+		"prettier": "^2.7.1",
+		"typescript": "^4.8.4",
+		"vite": "^3.2.4",
+		"vitest": "^0.25.2"
+	},
+	"release": {
+		"branches": [
+			"master",
+			{
+				"name": "beta",
+				"prerelease": true
+			},
+			{
+				"name": "alpha",
+				"prerelease": true
+			}
+		]
+	},
+	"lint-staged": {
+		"*.{js,jsx,css,ts,tsx}": [
+			"prettier --write",
+			"eslint"
+		],
+		"*.md": [
+			"prettier --write"
+		]
+	},
+	"commitlint": {
+		"extends": [
+			"@commitlint/config-conventional"
+		]
+	},
+	"eslintConfig": {
+		"env": {
+			"browser": true,
+			"es2021": true,
+			"node": true
+		},
+		"extends": [
+			"eslint:recommended",
+			"plugin:@typescript-eslint/recommended",
+			"prettier"
+		],
+		"parser": "@typescript-eslint/parser",
+		"parserOptions": {
+			"ecmaVersion": "latest",
+			"sourceType": "module"
+		},
+		"plugins": [
+			"@typescript-eslint"
+		],
+		"rules": {
+			"@typescript-eslint/no-explicit-any": "off",
+			"@typescript-eslint/no-var-requires": 1,
+			"no-async-promise-executor": 0,
+			"no-undef": 0
+		}
+	},
+	"prettier": {
+		"semi": false,
+		"singleQuote": true,
+		"tabWidth": 2,
+		"trailingComma": "es5",
+		"useTabs": true
+	}
 }
