@@ -1,5 +1,6 @@
 /* eslint-disable no-undef */
-import StoryblokClient from '../dist/index'
+import { expect, test, describe } from 'vitest'
+import StoryblokClient from '../'
 
 const accessToken = 'w0yFvs04aKF2rpz6F8OfIQtt'
 const cache = {
@@ -15,7 +16,7 @@ describe('Client should accept response interceptor as a function', () => {
 			return res
 		},
 	})
-	it('should RESPONSE function DO EXIST', async () => {
+	test('should RESPONSE function DO EXIST', async () => {
 		await Storyblok.getAll('cdn/links')
 		expect(Storyblok.client.responseInterceptor).toBeTruthy()
 	})
@@ -26,7 +27,7 @@ describe('Client should be initialized without interceptors', () => {
 		accessToken,
 		cache,
 	})
-	it('should RESPONSE function DO NOT EXIST', async () => {
+	test('should RESPONSE function DO NOT EXIST', async () => {
 		await Storyblok.getAll('cdn/links')
 		expect(Storyblok.client.responseInterceptor).toBeFalsy()
 	})
